refactor(github-issues-runtime-tests): extract core parsing module (#1266)
